+++ tests/components/index.js
@@ -1,0 +1,2 @@
+export TestComponent from './component'
+export TinyComponent from './tiny-component'
