+++ tests/components/tiny-component.jsx
