+++ tests/no-dom.jsx
@@ -4,7 +4,7 @@ import Test from '../src/no-dom'
 import { Find } from '../src/middleware'
 import { expect } from 'chai';
 
-import TestComponent from './component'
+import { TestComponent } from './components'
 
 describe('no dom', () => {
