+++ tests/shallowRender.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 
 import Test from '../src/legit-tests'
-import {Find} from '../src/middleware'
+import { Find } from '../src/middleware'
 import { expect } from 'chai';
 
-import TestComponent from './component'
+import { TestComponent } from './components'
 
 describe('Shallow Render', () => {
